task: improve loading of messages
Veged has found an edgecase: if messages loaded at first load would fit
into first screen, no scrollbar would appear, so it would not be
possible to load more messages.

This commit fixes it and also improves the loading of messages.
Also the number of loaded messages is extracted to constants.